Add integration tests with Redis (#11)
* Setting up a Redis instance in the integration tests github workflow
* Making an integration test for end-to-end training
