+++ .github/workflows/docs_build.yml
@@ -31,7 +31,7 @@ jobs:
           source $(poetry env info --path)/bin/activate
           poetry install --with docs,test
           cd docs && rm -rf source/reference/api/_autosummary && make html
-          cd .. && coverage run -m pytest -m "not integration_test" && coverage xml && coverage report -m
+          cd .. && coverage run -m pytest florist/tests/unit && coverage xml && coverage report -m
 #      - name: Upload coverage to Codecov
 #        uses: Wandalen/wretry.action@v1.4.4
 #        with:
