+++ .github/workflows/integration_tests.yaml
@@ -42,6 +42,10 @@ jobs:
       - uses: actions/setup-python@v5.0.0
         with:
           python-version: '3.9'
+      - name: Setup redis
+        uses: supercharge/redis-github-action@1.2.0
+        with:
+          redis-version: 7.2.4
       - name: Install dependencies and check code
         run: |
           poetry env use '3.9'
