+++ florist/tests/integration/api/launchers/test_launch.py
@@ -1,42 +1,20 @@
 import os
 import re
 import tempfile
-from functools import partial
 from pathlib import Path
-from typing import Callable, Dict
 
 import torch
-from fl4health.server.base_server import FlServer
 
 from florist.api.launchers.local import launch
-from florist.api.clients.mnist import MnistClient, MnistNet
-from florist.tests.utils.api.fl4health_utils import get_server_fedavg
-
-
-def fit_config(batch_size: int, local_epochs: int, current_server_round: int) -> Dict[str, int]:
-    return {
-        "batch_size": batch_size,
-        "current_server_round": current_server_round,
-        "local_epochs": local_epochs,
-    }
-
-
-def get_server(
-    fit_config: Callable[..., Dict[str, int]] = fit_config,
-    n_clients: int = 2,
-    batch_size: int = 8,
-    local_epochs: int = 1,
-) -> FlServer:
-    fit_config_fn = partial(fit_config, batch_size, local_epochs)
-    server = get_server_fedavg(model=MnistNet(), n_clients=n_clients, fit_config_fn=fit_config_fn)
-    return server
+from florist.api.clients.mnist import MnistClient
+from florist.tests.utils.api.launch_utils import get_server
 
 
 def assert_string_in_file(file_path: str, search_string: str) -> bool:
     with open(file_path, "r") as f:
         file_contents = f.read()
         match = re.search(search_string, file_contents)
-        return match is not None
+        assert match is not None
 
 
 def test_launch() -> None:
